\name{cet}
\alias{cet}
\docType{data}
\title{
Cetaceos
}
\description{
Cetaceos.
}
\usage{data("cet")}
\format{
  A data frame with 180 observations on the following 6 variables.
  \describe{
    \item{\code{DolphinID}}{a numeric vector}
    \item{\code{Species}}{a factor with levels \code{Delphinusdelphis} \code{Lagenorhynchusacutus} \code{Phocoenaphocoena} \code{Stenellacoeruleoalba} \code{Stenellafrontalis} \code{Tursiopstruncatus}}
    \item{\code{Age}}{a numeric vector}
    \item{\code{Sex}}{a numeric vector}
    \item{\code{Stain}}{a factor with levels \code{Elrich} \code{Mayer} \code{Toluidine}}
    \item{\code{Location}}{a factor with levels \code{Scotland} \code{Spain}}
  }
}

\examples{
data(cet)

}
\keyword{datasets}
