\name{fruit}
\alias{fruit}
\docType{data}
\title{
Fruit
}
\description{
Fruit.
}
\usage{data("fruit")}
\format{
  A data frame with 428 observations on the following 4 variables.
  \describe{
    \item{\code{TAG}}{a numeric vector}
    \item{\code{DBH}}{a numeric vector}
    \item{\code{FRUIT}}{a numeric vector}
    \item{\code{YEAR}}{a numeric vector}
  }
}

\examples{
data(fruit)

}
\keyword{datasets}
