\name{sn}
\alias{sn}
\docType{data}
\title{
Sn
}
\description{
Sn.
}
\usage{data("sn")}
\format{
  A data frame with 72 observations on the following 4 variables.
  \describe{
    \item{\code{Habitat}}{a factor with levels \code{Forest} \code{Grassland} \code{Shrubland}}
    \item{\code{Plot}}{a numeric vector}
    \item{\code{Scenario}}{a factor with levels \code{Control} \code{Dry Summer} \code{Wet Summer}}
    \item{\code{Total.Biomass}}{a numeric vector}
  }
}
\examples{
data(sn)
}
\keyword{datasets}
